zania/requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==5.1.7
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
factory_boy==3.3.3
Faker==37.0.0
gunicorn==23.0.0
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.1
